# server/api/windrunner-app/src/main/java/com/windrunner/server/external/v1/api/ExternalProjectController.java
package com.windrunner.server.external.v1.api;

import com.windrunner.server.api.ApiResponse;
import com.windrunner.server.apikey.ApiKeyScopes;
import com.windrunner.server.audit.*;
import com.windrunner.server.auth.security.AppRoles;
import com.windrunner.server.external.auth.ExternalAccessService;
import com.windrunner.server.external.v1.dto.ExternalProjectMemberResponse;
import com.windrunner.server.external.v1.dto.ExternalProjectResponse;
import com.windrunner.server.external.v1.dto.ExternalProjectTeamResponse;
import com.windrunner.server.id.EntityIdGenerator;
import com.windrunner.server.id.EntityIdType;
import com.windrunner.server.project.ProjectAccessService;
import com.windrunner.server.project.ProjectContentDeletionService;
import com.windrunner.server.project.ProjectRoles;
import com.windrunner.server.project.api.CreateProjectRequest;
import com.windrunner.server.project.domain.Project;
import com.windrunner.server.project.domain.ProjectMember;
import com.windrunner.server.project.persistence.ProjectMemberRepository;
import com.windrunner.server.project.persistence.ProjectRepository;
import com.windrunner.server.team.api.TeamLinkRequest;
import com.windrunner.server.team.domain.ProjectTeam;
import com.windrunner.server.team.persistence.ProjectTeamRepository;
import com.windrunner.server.team.persistence.TeamRepository;
import com.windrunner.server.user.domain.AppUser;
import com.windrunner.server.user.persistence.AppUserRepository;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.servlet.http.HttpServletRequest;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.server.ResponseStatusException;

import java.util.*;

@RequiredArgsConstructor
@RestController
@RequestMapping("/api/v1/projects")
@Tag(name = "Projects", description = "Create projects, manage members and linked teams.")
public class ExternalProjectController {

    private final ProjectRepository projectRepository;
    private final ProjectMemberRepository projectMemberRepository;
    private final ProjectTeamRepository projectTeamRepository;
    private final TeamRepository teamRepository;
    private final AppUserRepository appUserRepository;
    private final ProjectAccessService projectAccessService;
    private final AuditLogService auditLogService;
    private final ExternalAccessService externalAccessService;
    private final EntityIdGenerator idGenerator;
    private final ProjectContentDeletionService projectContentDeletionService;

    @GetMapping
    public ApiResponse<List<ExternalProjectResponse>> listProjects(@RequestParam(name = "page", defaultValue = "0") int page,
                                                   @RequestParam(name = "size", defaultValue = "50") int size,
                                                   HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECTS_READ);
        int normalizedPage = Math.max(page, 0);
        int normalizedSize = Math.max(1, Math.min(size, 100));
        long totalItems;
        List<Project> items;
        if (AppRoles.isSuperAdmin(actor.getGlobalRole())) {
            totalItems = projectRepository.countProjects();
            items = projectRepository.findAllPage(normalizedSize, (long) normalizedPage * normalizedSize);
        } else {
            totalItems = projectRepository.countVisibleToUser(actor.getId());
            items = projectRepository.findVisibleToUserPaged(actor.getId(), normalizedSize, (long) normalizedPage * normalizedSize);
        }
        return ApiResponse.page(
                items.stream().map(ExternalProjectResponse::from).toList(),
                normalizedPage,
                normalizedSize,
                totalItems,
                (int) Math.ceil(totalItems / (double) normalizedSize));
    }

    @GetMapping("/{id}")
    public ApiResponse<ExternalProjectResponse> getProject(@PathVariable("id") String id, HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECTS_READ);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.VIEWER);
        return ApiResponse.success(ExternalProjectResponse.from(requireProject(id)));
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    @Transactional
    public ApiResponse<ExternalProjectResponse> createProject(@RequestBody CreateProjectRequest createRequest, HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECTS_WRITE);
        if (createRequest == null) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Request body is required");
        }
        String name = ExternalInputValidation.requiredText(createRequest.name(), "Project name",
                ExternalInputValidation.MAX_NAME_LENGTH);
        List<String> ownerUserIds = normalizeIds(createRequest.ownerUserIds(), "Owner user id is required");
        List<String> ownerTeamIds = normalizeIds(createRequest.ownerTeamIds(), "Owner team id is required");
        if (ownerUserIds.isEmpty() && ownerTeamIds.isEmpty()) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "At least one project owner is required");
        }
        for (String ownerUserId : ownerUserIds) {
            if (!appUserRepository.existsById(ownerUserId)) {
                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Project owner user not found");
            }
        }
        for (String ownerTeamId : ownerTeamIds) {
            if (!teamRepository.existsById(ownerTeamId)) {
                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Project owner team not found");
            }
        }

        Project project = new Project();
        project.setId(idGenerator.generate(EntityIdType.PROJECT));
        project.setName(name);
        project.setCreatedByUserId(actor.getId());
        projectRepository.insert(project.getId(), project.getName(), actor.getId());
        // The API-key owner always becomes an owner so the caller retains access to the project it created.
        projectMemberRepository.upsert(project.getId(), actor.getId(), ProjectRoles.OWNER);
        for (String ownerUserId : ownerUserIds) {
            projectMemberRepository.upsert(project.getId(), ownerUserId, ProjectRoles.OWNER);
        }
        for (String ownerTeamId : ownerTeamIds) {
            projectTeamRepository.upsert(project.getId(), ownerTeamId, ProjectRoles.OWNER);
        }
        auditLogService.logAfterCommit(new AuditLogEntry(
                actor.getId(),
                AuditActions.CREATE,
                AuditEntityTypes.PROJECT,
                project.getId(),
                project.getId(),
                AuditOutcomes.SUCCESS,
                "Created project " + project.getName(),
                null,
                auditLogService.json(projectSnapshot(project)),
                null,
                null));
        return ApiResponse.success(ExternalProjectResponse.from(requireProject(project.getId())));
    }

    @PutMapping("/{id}")
    @Transactional
    public ApiResponse<ExternalProjectResponse> updateProject(@PathVariable("id") String id,
                                              @RequestBody Project project,
                                              HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECTS_WRITE);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.OWNER);
        Project beforeProject = requireProject(id);
        Map<String, Object> before = projectSnapshot(beforeProject);
        validateName(project);
        project.setId(id);
        if (projectRepository.update(project.getId(), project.getName()) == 0) {
            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Project not found");
        }
        Map<String, Object> after = projectSnapshot(project);
        auditLogService.logAfterCommit(new AuditLogEntry(
                actor.getId(),
                AuditActions.UPDATE,
                AuditEntityTypes.PROJECT,
                project.getId(),
                project.getId(),
                AuditOutcomes.SUCCESS,
                "Updated project " + project.getName(),
                auditLogService.json(before),
                auditLogService.json(after),
                auditLogService.changes(before, after),
                null));
        return ApiResponse.success(ExternalProjectResponse.from(requireProject(id)));
    }

    @DeleteMapping("/{id}")
    @Transactional
    public ApiResponse<Void> deleteProject(@PathVariable("id") String id, HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECTS_WRITE);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.OWNER);
        Project project = requireProject(id);
        Map<String, Object> before = projectSnapshot(project);
        projectContentDeletionService.deleteProjectContent(id);
        projectTeamRepository.deleteByProjectId(id);
        projectMemberRepository.deleteByProjectId(id);
        projectRepository.deleteById(id);
        auditLogService.logAfterCommit(new AuditLogEntry(
                actor.getId(),
                AuditActions.DELETE,
                AuditEntityTypes.PROJECT,
                project.getId(),
                project.getId(),
                AuditOutcomes.SUCCESS,
                "Deleted project " + project.getName(),
                auditLogService.json(before),
                null,
                null,
                null));
        return ApiResponse.success();
    }

    @GetMapping("/{id}/teams")
    public ApiResponse<List<ExternalProjectTeamResponse>> listProjectTeams(@PathVariable("id") String id,
                                                                            @RequestParam(name = "page", defaultValue = "0") int page,
                                                                            @RequestParam(name = "size", defaultValue = "50") int size,
                                                                            HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECT_ACCESS_READ);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.VIEWER);
        requireProject(id);
        int normalizedPage = Math.max(page, 0);
        int normalizedSize = Math.max(1, Math.min(size, 100));
        long totalItems = projectTeamRepository.countByProjectId(id);
        return ApiResponse.page(projectTeamRepository.findPageByProjectId(id, normalizedSize,
                        (long) normalizedPage * normalizedSize).stream()
                .map(ExternalProjectTeamResponse::from)
                .toList(), normalizedPage, normalizedSize, totalItems,
                (int) Math.ceil(totalItems / (double) normalizedSize));
    }

    @GetMapping("/{id}/members")
    public ApiResponse<List<ExternalProjectMemberResponse>> listProjectMembers(@PathVariable("id") String id,
                                                                                @RequestParam(name = "page", defaultValue = "0") int page,
                                                                                @RequestParam(name = "size", defaultValue = "50") int size,
                                                                                HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECT_ACCESS_READ);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.VIEWER);
        requireProject(id);
        int normalizedPage = Math.max(page, 0);
        int normalizedSize = Math.max(1, Math.min(size, 100));
        long totalItems = projectMemberRepository.countByProjectId(id);
        return ApiResponse.page(projectMemberRepository.findPageByProjectId(id, normalizedSize,
                        (long) normalizedPage * normalizedSize).stream()
                .map(ExternalProjectMemberResponse::from)
                .toList(), normalizedPage, normalizedSize, totalItems,
                (int) Math.ceil(totalItems / (double) normalizedSize));
    }

    @PostMapping("/{id}/members")
    @ResponseStatus(HttpStatus.CREATED)
    @Transactional
    public ApiResponse<ExternalProjectMemberResponse> upsertProjectMember(@PathVariable("id") String id,
                                                          @RequestBody TeamLinkRequest linkRequest,
                                                          HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECT_ACCESS_WRITE);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.OWNER);
        Project project = requireProject(id);
        String userId = requireText(linkRequest == null ? null : linkRequest.userId(), "User id is required");
        if (!appUserRepository.existsById(userId)) {
            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
        }
        String role = normalizeProjectRole(linkRequest == null ? null : linkRequest.role());
        ProjectMember existing = projectMemberRepository.findByProjectIdAndUserId(id, userId).orElse(null);
        projectAccessService.requireAnotherOwnerBeforeRemovingOwner(
                id,
                existing != null && ProjectRoles.OWNER.equals(existing.getRole()) && !ProjectRoles.OWNER.equals(role));
        projectMemberRepository.upsert(id, userId, role);

        ProjectMember projectMember = new ProjectMember();
        projectMember.setProjectId(id);
        projectMember.setUserId(userId);
        projectMember.setRole(role);
        auditLogService.logAfterCommit(new AuditLogEntry(
                actor.getId(),
                AuditActions.UPDATE,
                AuditEntityTypes.PROJECT,
                id,
                id,
                AuditOutcomes.SUCCESS,
                "Updated project member for " + project.getName(),
                null,
                null,
                null,
                auditLogService.json(Map.of("operation", "UPSERT_MEMBER", "userId", userId, "role", role))));
        return ApiResponse.success(ExternalProjectMemberResponse.from(projectMember));
    }

    @DeleteMapping("/{id}/members/{userId}")
    @Transactional
    public ApiResponse<Void> removeProjectMember(@PathVariable("id") String id,
                                                 @PathVariable("userId") String userId,
                                                 HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECT_ACCESS_WRITE);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.OWNER);
        Project project = requireProject(id);
        ProjectMember existing = projectMemberRepository.findByProjectIdAndUserId(id, userId)
                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Project member not found"));
        projectAccessService.requireAnotherOwnerBeforeRemovingOwner(id, ProjectRoles.OWNER.equals(existing.getRole()));
        projectMemberRepository.delete(id, userId);
        auditLogService.logAfterCommit(new AuditLogEntry(
                actor.getId(),
                AuditActions.UPDATE,
                AuditEntityTypes.PROJECT,
                id,
                id,
                AuditOutcomes.SUCCESS,
                "Removed project member from " + project.getName(),
                null,
                null,
                null,
                auditLogService.json(Map.of("operation", "REMOVE_MEMBER", "userId", userId))));
        return ApiResponse.success();
    }

    @PostMapping("/{id}/teams")
    @ResponseStatus(HttpStatus.CREATED)
    @Transactional
    public ApiResponse<ExternalProjectTeamResponse> assignTeam(@PathVariable("id") String id,
                                               @RequestBody TeamLinkRequest linkRequest,
                                               HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECT_ACCESS_WRITE);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.OWNER);
        Project project = requireProject(id);
        String teamId = requireText(linkRequest == null ? null : linkRequest.teamId(), "Team id is required");
        if (!teamRepository.existsById(teamId)) {
            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Team not found");
        }
        String role = normalizeProjectRole(linkRequest == null ? null : linkRequest.role());
        ProjectTeam existing = projectTeamRepository.findByProjectIdAndTeamId(id, teamId).orElse(null);
        projectAccessService.requireAnotherOwnerBeforeRemovingOwner(
                id,
                existing != null && ProjectRoles.OWNER.equals(existing.getRole()) && !ProjectRoles.OWNER.equals(role));
        projectTeamRepository.upsert(id, teamId, role);

        ProjectTeam projectTeam = new ProjectTeam();
        projectTeam.setProjectId(id);
        projectTeam.setTeamId(teamId);
        projectTeam.setRole(role);
        auditLogService.logAfterCommit(new AuditLogEntry(
                actor.getId(),
                AuditActions.UPDATE,
                AuditEntityTypes.PROJECT,
                id,
                id,
                AuditOutcomes.SUCCESS,
                "Assigned team to project " + project.getName(),
                null,
                null,
                null,
                auditLogService.json(Map.of("operation", "ASSIGN_TEAM", "teamId", teamId, "role", role))));
        return ApiResponse.success(ExternalProjectTeamResponse.from(projectTeam));
    }

    @DeleteMapping("/{id}/teams/{teamId}")
    @Transactional
    public ApiResponse<Void> unassignTeam(@PathVariable("id") String id,
                                          @PathVariable("teamId") String teamId,
                                          HttpServletRequest request) {
        AppUser actor = externalAccessService.requireScope(request, ApiKeyScopes.PROJECT_ACCESS_WRITE);
        projectAccessService.requireProjectRole(id, actor, ProjectRoles.OWNER);
        Project project = requireProject(id);
        ProjectTeam existing = projectTeamRepository.findByProjectIdAndTeamId(id, teamId)
                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Project team not found"));
        projectAccessService.requireAnotherOwnerBeforeRemovingOwner(id, ProjectRoles.OWNER.equals(existing.getRole()));
        projectTeamRepository.delete(id, teamId);
        auditLogService.logAfterCommit(new AuditLogEntry(
                actor.getId(),
                AuditActions.UPDATE,
                AuditEntityTypes.PROJECT,
                id,
                id,
                AuditOutcomes.SUCCESS,
                "Unassigned team from project " + project.getName(),
                null,
                null,
                null,
                auditLogService.json(Map.of("operation", "UNASSIGN_TEAM", "teamId", teamId))));
        return ApiResponse.success();
    }

    private void validateName(Project project) {
        if (project == null) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Request body is required");
        }
        project.setName(ExternalInputValidation.requiredText(project.getName(), "Project name",
                ExternalInputValidation.MAX_NAME_LENGTH));
    }

    private Map<String, Object> projectSnapshot(Project project) {
        Map<String, Object> snapshot = new LinkedHashMap<>();
        snapshot.put("id", project.getId());
        snapshot.put("name", project.getName());
        snapshot.put("createdByUserId", project.getCreatedByUserId());
        snapshot.put("archivedAt", project.getArchivedAt() == null ? null : project.getArchivedAt().toString());
        return snapshot;
    }

    private Project requireProject(String id) {
        return projectRepository.findById(id)
                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Project not found"));
    }

    private String requireText(String value, String message) {
        return ExternalInputValidation.requiredText(value, message, ExternalInputValidation.MAX_ID_LENGTH);
    }

    private List<String> normalizeIds(List<String> values, String blankMessage) {
        if (values == null || values.isEmpty()) {
            return List.of();
        }
        ExternalInputValidation.requireMaxSize(values, "Owner ids", ExternalInputValidation.MAX_ID_LIST_SIZE);
        Set<String> normalized = new LinkedHashSet<>();
        for (String value : values) {
            if (value == null || value.isBlank()) {
                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, blankMessage);
            }
            normalized.add(ExternalInputValidation.requiredText(value, blankMessage,
                    ExternalInputValidation.MAX_ID_LENGTH));
        }
        return List.copyOf(normalized);
    }

    private String normalizeProjectRole(String role) {
        try {
            return ProjectRoles.normalize(role);
        } catch (IllegalArgumentException e) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
        }
    }
}